Don't use Arc for eh_frame_hdr data.
We only need to use Arc for eh_frame data, because the UnwindContext in the
UnwinderCache stores gimli Readers, so that cache needs shared ownership of
the eh_frame data. Well, it doesn't really need ownership, it would be fine
if the cache stored some kind of weak pointer, because we don't plan on accessing
any cached stuff for modules that have been unloaded, but then you'd need to
decide what to do if you do encounter any dead pointers, and that sounds hard.