Change the Module creation interface.
There is now a more general `ModuleSectionInfo` trait which is used to
query sections/segments as needed. This moves a lot of cumbersome
conditional code into the library to make clients' lives easier.

This also adds `object` as an optional dependency, where (if enabled)
the trait is implemented for references to `object::read::Object` types.