Add basic is_first_frame support for PE.
Previously, the PE backend always attempted to perform instruction
analysis. This was not only inefficient but also wrong. `address` was
subtracted by one for return addresses found on the stack, which makes
the lookup point to the previous instruction, but also meant that
instruction decoding shouldn't happen, as it will likely be reading from
the middle of an instruction.

Implement basic is_first_frame support, similar to what we are doing for
DWARF. Later, we should extend this so that PUSH_MACHFRAME and
.cfi_signal_frame would also be considered a "first frame".