Make Unwinder Clone-eable.
This makes it easier for samply to handle forked processes correctly.
When a process forks, the mappings in the old process are carried over
into the new process. We want to be able to unwind stacks from the new
process without manually replaying the lib mapping operations to the
unwinder.